Move fromMultiScanFolder into dedicated MRMultiScanLoad.h/.cpp
Relocate the function together with its file-name parsing and .pose reading
helpers out of MRPointsLoad into new MRMultiScanLoad.{h,cpp}, and register both
files in MRMesh.vcxproj. No behavior change.
